pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simple-poller"
version = "0.1.0"
description = ""
authors = [{ name = "Chuancong Gao", email = "chuanconggao@users.noreply.github.com" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = []

[project.optional-dependencies]
sqs = [
    "boto3>=1.37.13",
]

[dependency-groups]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ipython>=9.0.2",
    "mypy>=1.7.1",
]

[tool.coverage.run]
source = [
    "src",
]

[tool.pytest.ini_options]
addopts = "--cov --cov-report xml:cov.xml --cov-report html --cov-report term-missing"
pythonpath = [
    "src",
]
